requirements.txt
# Core MCP dependencies
mcp[cli]>=1.0.0

# Web UI
streamlit>=1.28.0

# AI Integration
anthropic>=0.18.0
google-generativeai>=0.3.0

# Image processing
Pillow>=10.0.0

# Numerical operations
numpy>=1.24.0

# Development tools
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
